Move input-related patches to a separate file
