+++ MWSE/PatchInput.h
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace mwse::patch::input {
+	void install();
+	void installPostLua();
+}
