+++ MWSE/PatchUtil.cpp
@@ -1,6 +1,7 @@
 #include "PatchUtil.h"
 
 #include "PatchAudio.h"
+#include "PatchInput.h"
 #include "PatchMagic.h"
 #include "PatchMWScript.h"
 #include "PatchNI.h"
@@ -22,7 +23,6 @@
 #include "TES3Game.h"
 #include "TES3GameFile.h"
 #include "TES3GameSetting.h"
-#include "TES3InputController.h"
 #include "TES3ItemData.h"
 #include "TES3Light.h"
 #include "TES3LoadScreenManager.h"
@@ -195,45 +195,6 @@ namespace mwse::patch {
 
 	UINT WINAPI	OverrideDontThreadLoad(LPCSTR lpAppName, LPCSTR lpKeyName, INT nDefault, LPCSTR lpFileName) {
 		return TES3::DataHandler::suppressThreadLoad || TES3::DataHandler::dontThreadLoad;
-	}
-
-	//
-	// Patch: Make Morrowind believe that it is always the front window in the main gameplay loop block.
-	//
-
-	static HWND lastForegroundWindow = 0;
-	HWND __stdcall PatchGetActiveWindowForMainLoop() {
-		const auto worldController = TES3::WorldController::get();
-		const auto mainWindowHandle = worldController->Win32_hWndParent;
-
-		const auto foregroundWindow = GetForegroundWindow();
-		if (foregroundWindow != mainWindowHandle && foregroundWindow != lastForegroundWindow) {
-			auto inputController = worldController->inputController;
-			if (inputController) {
-				inputController->clearTransientInputState();
-				if (inputController->mouse) {
-					inputController->mouse->Unacquire();
-				}
-				if (inputController->keyboard) {
-					inputController->keyboard->Unacquire();
-				}
-			}
-		}
-
-		lastForegroundWindow = foregroundWindow;
-
-		if (Configuration::RunInBackground) {
-			return mainWindowHandle;
-		}
-		return GetActiveWindow();
-	}
-
-	void __fastcall PatchGetMorrowindMainWindow_NoBackgroundInput(TES3::InputController* inputController) {
-		if (GetForegroundWindow() != TES3::WorldController::get()->Win32_hWndParent) {
-			return;
-		}
-
-		inputController->readKeyState();
 	}
 
 	//
@@ -941,38 +902,6 @@ namespace mwse::patch {
 	}
 
 	//
-	// Patch: Expand keyboard key translations
-	//
-
-	inline static void WritePatchKeyCharacter(unsigned int key, char character) {
-		se::memory::writeByteUnprotected(0x775148 + key, character); // US, Unshifted
-		se::memory::writeByteUnprotected(0x775248 + key, character); // US, Shifted
-		se::memory::writeValueEnforced<char>(0x775348 + key, 0, character); // DE, Unshifted
-		se::memory::writeValueEnforced<char>(0x775448 + key, 0, character); // DE, Shifted
-		se::memory::writeValueEnforced<char>(0x775548 + key, 0, character); // FR, Unshifted
-		se::memory::writeValueEnforced<char>(0x775648 + key, 0, character); // FR, Shifted
-	}
-
-	static void PatchExpandKeyboardCharacterTranslations() {
-		WritePatchKeyCharacter(DIK_NUMPAD0, '0');
-		WritePatchKeyCharacter(DIK_NUMPAD1, '1');
-		WritePatchKeyCharacter(DIK_NUMPAD2, '2');
-		WritePatchKeyCharacter(DIK_NUMPAD3, '3');
-		WritePatchKeyCharacter(DIK_NUMPAD4, '4');
-		WritePatchKeyCharacter(DIK_NUMPAD5, '5');
-		WritePatchKeyCharacter(DIK_NUMPAD6, '6');
-		WritePatchKeyCharacter(DIK_NUMPAD7, '7');
-		WritePatchKeyCharacter(DIK_NUMPAD8, '8');
-		WritePatchKeyCharacter(DIK_NUMPAD9, '9');
-		WritePatchKeyCharacter(DIK_NUMPADEQUALS, '=');
-		WritePatchKeyCharacter(DIK_NUMPADMINUS, '-');
-		WritePatchKeyCharacter(DIK_NUMPADPERIOD, '.');
-		WritePatchKeyCharacter(DIK_NUMPADPLUS, '+');
-		WritePatchKeyCharacter(DIK_NUMPADSLASH, '/');
-		WritePatchKeyCharacter(DIK_NUMPADSTAR, '*');
-	}
-
-	//
 	// Patch: Fix crash when trying to unequip a nocked projectile item while still using an item index for its position.
 	// 
 	// Before serializing, the nocked projectile is converted into an item index. Some mods may try to unequip the index
@@ -1147,100 +1076,6 @@ namespace mwse::patch {
 		for (auto& info : other->info) {
 			PatchMergeDialogueInfo(self, NULL, info, false);
 		}
-	}
-
-	//
-	// Patch: Be better about showing/hiding the cursor.
-	//
-
-	static WNDPROC originalWindowProc = nullptr;
-	using gCursorShown = se::memory::ExternalGlobal<bool, 0x776D0C>;
-	static bool showCursorFlag = true;
-
-	static void SetShowCursorState(bool shown, bool force = false) {
-		if (!force && showCursorFlag == shown) {
-			return;
-		}
-
-		if (shown) {
-			while (ShowCursor(TRUE) < 0);
-		}
-		else {
-			while (ShowCursor(FALSE) >= 0);
-		}
-		showCursorFlag = shown;
-	}
-
-	static void SetCursorShown(HWND hWnd, bool shown, bool forceShow = false) {
-		gCursorShown::set(shown);
-		SetShowCursorState(shown, forceShow && shown);
-
-		const auto worldController = TES3::WorldController::get();
-		if (!worldController) {
-			return;
-		}
-
-		const auto inputController = worldController->inputController;
-		if (!inputController) {
-			return;
-		}
-
-		// Sync mouse state.
-		if (inputController->mouse) {
-			if (shown) {
-				inputController->mouse->Unacquire();
-			}
-			else if (GetForegroundWindow() == hWnd) {
-				inputController->mouse->Acquire();
-			}
-		}
-
-		// Sync keyboard state.
-		if (inputController->keyboard) {
-			if (shown) {
-				inputController->keyboard->Unacquire();
-			}
-			else if (GetForegroundWindow() == hWnd) {
-				inputController->keyboard->Acquire();
-			}
-		}
-	}
-
-	static void PatchWindProc_CursorHitTest(se::windows::DialogProcContext& context) {
-		context.callOriginalFunction();
-		const auto hWnd = context.getWindowHandle();
-		const auto result = context.getResult();
-		auto shouldShow = result != HTCLIENT;
-
-		SetCursorShown(hWnd, shouldShow || GetForegroundWindow() != hWnd);
-	}
-
-	static LRESULT __stdcall PatchWindProc(HWND hWnd, UINT msg, WPARAM wParam, LPARAM lParam) {
-		se::windows::DialogProcContext context(hWnd, msg, wParam, lParam, (DWORD)originalWindowProc);
-
-		switch (msg) {
-		case WM_ACTIVATE:
-			SetCursorShown(hWnd, context.getLOWParam() == WA_INACTIVE, true);
-			break;
-		case WM_ACTIVATEAPP:
-			SetCursorShown(hWnd, !wParam, true);
-			break;
-		case WM_SETFOCUS:
-			SetCursorShown(hWnd, false);
-			break;
-		case WM_KILLFOCUS:
-			SetCursorShown(hWnd, true, true);
-			break;
-		case WM_NCHITTEST:
-			PatchWindProc_CursorHitTest(context);
-			break;
-		}
-
-		if (!context.hasResult()) {
-			context.callOriginalFunction();
-		}
-
-		return context.getResult();
 	}
 
 	// Replacement for the Sleep() call inside DataHandler::sub_48F5F0's
@@ -1552,6 +1387,7 @@ namespace mwse::patch {
 		using se::memory::writeDoubleWordEnforced;
 
 		audio::install();
+		input::install();
 		magic::install();
 		mwscript::install();
 		ni::install();
@@ -1976,9 +1812,6 @@ namespace mwse::patch {
 		// Patch: Fix missing nullptr check when determining object types.
 		DoPatchSafeGetObjectType();
 
-		// Patch: Expand keyboard key translations
-		PatchExpandKeyboardCharacterTranslations();
-
 		// Patch: Fix crash when trying to unequip a nocked projectile item while still using an item index for its position.
 		genNOPUnprotected(0x4968E1, 0x4968FB - 0x4968E1);
 		writePatchCodeUnprotected(0x4968E1, (BYTE*)&PatchUnequipIndexedProjectileSetup, PatchUnequipIndexedProjectileSetup_size);
@@ -2010,11 +1843,6 @@ namespace mwse::patch {
 		genCallEnforced(0x4EF99F, 0x4EEFC0, *reinterpret_cast<DWORD*>(&PhysicalObject_createBoundingBox));
 		genCallEnforced(0x4EFE70, 0x4EEFC0, *reinterpret_cast<DWORD*>(&PhysicalObject_createBoundingBox));
 
-		// Patch: Store last read key state.
-		auto InputController_readButtonPressed = &TES3::InputController::readButtonPressed;
-		genCallEnforced(0x58E8C6, 0x406950, *reinterpret_cast<DWORD*>(&InputController_readButtonPressed));
-		genCallEnforced(0x5BCA1D, 0x406950, *reinterpret_cast<DWORD*>(&InputController_readButtonPressed));
-
 		// Patch: Optimize map rendering when crossing cell borders.
 		genJumpEnforced(0x420415, 0x5E99C0, reinterpret_cast<DWORD>(&PatchOptimizeMapUpdates_UpdateMapRenderCached)); // MapController::updateMapRender tail-jmp
 		genCallEnforced(0x5E9757, 0x5E99C0, reinterpret_cast<DWORD>(&PatchOptimizeMapUpdates_UpdateMapRenderCached)); // ui_showMapMenu
@@ -2045,30 +1873,8 @@ namespace mwse::patch {
 		using se::memory::genPushEnforced;
 
 		audio::installPostLua();
+		input::installPostLua();
 		ni::installPostLua();
-
-		// Patch: Be better about showing/hiding the cursor.
-		originalWindowProc = (WNDPROC)SetWindowLongPtr(TES3::WorldController::get()->Win32_hWndParent, GWLP_WNDPROC, (LONG_PTR)PatchWindProc);
-
-		// Patch: Reset input state when focus changes outside Morrowind, even if it happened while paused in a debugger.
-		genCallUnprotected(0x41AB7D, reinterpret_cast<DWORD>(PatchGetActiveWindowForMainLoop), 0x6);
-
-		// Patch: The window is never out of focus.
-		if (Configuration::RunInBackground) {
-			writeByteUnprotected(0x416BC3 + 0x2 + 0x4, 1);
-			genCallEnforced(0x425313, 0x4065E0, reinterpret_cast<DWORD>(PatchGetMorrowindMainWindow_NoBackgroundInput));
-			genCallEnforced(0x4772CE, 0x4065E0, reinterpret_cast<DWORD>(PatchGetMorrowindMainWindow_NoBackgroundInput));
-			genCallEnforced(0x47798C, 0x4065E0, reinterpret_cast<DWORD>(PatchGetMorrowindMainWindow_NoBackgroundInput));
-			genCallEnforced(0x477E1E, 0x4065E0, reinterpret_cast<DWORD>(PatchGetMorrowindMainWindow_NoBackgroundInput));
-			genCallEnforced(0x5BC9E1, 0x4065E0, reinterpret_cast<DWORD>(PatchGetMorrowindMainWindow_NoBackgroundInput));
-			genCallEnforced(0x5BCA33, 0x4065E0, reinterpret_cast<DWORD>(PatchGetMorrowindMainWindow_NoBackgroundInput));
-		}
-
-		// Patch: Use non-exclusive keyboard cooperative levels and allow shell keys.
-		if (Configuration::NonExclusiveKeyboard) {
-			genPushEnforced(0x40627D, BYTE(DISCL_BACKGROUND | DISCL_NONEXCLUSIVE));
-			genPushEnforced(0x406291, BYTE(DISCL_FOREGROUND | DISCL_NONEXCLUSIVE));
-		}
 
 		// Patch: Replace the Sleep(100) call inside the background loader's
 		// progress-show polling loop with a wrapper that reads
